Update RangeSliderControl.cs
Corrected value types for the normalised min and max values. Currently when restoring state the values fail to be read with `java.lang.ClassCastException: java.lang.Double cannot be cast to java.lang.Float` and are reset to 0

+++ Xamarin.RangeSlider.Droid/RangeSliderControl.cs
@@ -937,8 +937,8 @@ namespace Xamarin.RangeSlider
         {
             var bundle = new Bundle();
             bundle.PutParcelable("SUPER", base.OnSaveInstanceState());
-            bundle.PutDouble("MIN", NormalizedMinValue);
-            bundle.PutDouble("MAX", NormalizedMaxValue);
+            bundle.PutFloat("MIN", NormalizedMinValue);
+            bundle.PutFloat("MAX", NormalizedMaxValue);
             return bundle;
         }
